requirements.txt
blinker==1.9.0
boto3==1.37.33
botocore==1.37.33
cassandra-driver==3.29.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
geomet==0.2.1.post1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
